+++ semcore/data-table/__tests__/data-table-accordion.browser-test.tsx
@@ -357,9 +357,20 @@ test.describe('Accordion in table', () => {
     });
   });
 
-  test('Verify table in table keyboard navigation', async ({ page }) => {
+  test('Verify table in table keyboard navigation when accordionMode=independent', async ({ page }) => {
+    let messages: string[] = [];
+    page.on('console', (msg) => {
+      if (msg.type() === 'log') {
+        messages.push(msg.text());
+      }
+    });
+
     const standPath = 'stories/components/data-table/docs/examples/table-in-table.tsx';
-    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    const htmlContent = await e2eStandToHtml(standPath, 'en', {
+      accordionMode: 'independent',
+      onAccordionToggle: 'function(type, rowIndex) { console.log("Accordion " + type + " for row #" + rowIndex); }',
+    });
 
     await page.setContent(htmlContent);
 
@@ -380,12 +391,200 @@ test.describe('Accordion in table', () => {
       await expect(cellinLastRow).toBeFocused();
     });
 
-    await test.step('Verify it is possible to scroll the last cell by keyboard when accordion expanded', async () => {
+    await test.step('Verify it is possible to scroll the last cell by keyboard when first accordion expanded', async () => {
+      await page.keyboard.press('ArrowUp');
       await page.keyboard.press('ArrowUp');
       await page.keyboard.press('ArrowUp');
       await page.keyboard.press('ArrowUp');
       await page.keyboard.press('Enter');
-      await page.waitForTimeout(250);
+      const accordionRow = page.locator('[aria-rowindex="5"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #0']);
+
+      for (let i = 0; i < 7; i++) await page.keyboard.press('ArrowDown');
+
+      const lastRow = page.locator('[data-ui-name="Body.Row"][aria-rowindex="9"]');
+      const cellinLastRow = lastRow.locator(
+        '[data-ui-name="Body.Cell"][aria-colindex="1"][data-aria-level="1"]',
+      );
+
+      await expect(cellinLastRow).toBeFocused();
+    });
+
+    await test.step('Verify it is possible to scroll the last cell by keyboard when second accordion expanded', async () => {
+      messages = [];
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('Enter');
+      const accordionRow = page.locator('[aria-rowindex="9"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #1']);
+      for (let i = 0; i < 6; i++) await page.keyboard.press('ArrowDown');
+
+      const lastRow = page.locator('[data-ui-name="Body.Row"][aria-rowindex="12"]');
+      const cellinLastRow = lastRow.locator(
+        '[data-ui-name="Body.Cell"][aria-colindex="1"][data-aria-level="1"]',
+      );
+
+      await expect(cellinLastRow).toBeFocused();
+    });
+
+    await test.step('Verify no issues with keyboard navigation up when second accordion just collapsed', async () => {
+      messages = [];
+
+      for (let i = 0; i < 6; i++) await page.keyboard.press('ArrowUp');
+      await expect(page.locator('[data-ui-name="ButtonLink"]').nth(1)).toBeFocused();
+      await page.keyboard.press('Enter');
+      const accordionRow = page.locator('[aria-rowindex="9"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'hidden' });
+      await page.waitForEvent('console', {
+        predicate: (msg) => msg.type() === 'log' && msg.text() === 'Accordion close for row #1',
+        timeout: 200,
+      });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion close for row #1']);
+      messages = [];
+
+      await page.keyboard.press('Enter');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #1']);
+
+      for (let i = 0; i < 4; i++) await page.keyboard.press('ArrowUp');
+
+      await expect(page.locator('[data-ui-name="ButtonLink"]').first()).toBeFocused();
+    });
+  });
+
+  test('Verify table in table mouse navigation when accordionMode=independent', async ({ page }) => {
+    let messages: string[] = [];
+    page.on('console', (msg) => {
+      if (msg.type() === 'log') {
+        messages.push(msg.text());
+      }
+    });
+
+    const standPath = 'stories/components/data-table/docs/examples/table-in-table.tsx';
+
+    const htmlContent = await e2eStandToHtml(standPath, 'en', {
+      accordionMode: 'independent',
+      onAccordionToggle: 'function(type, rowIndex) { console.log("Accordion " + type + " for row #" + rowIndex); }',
+    });
+
+    await page.setContent(htmlContent);
+
+    await test.step('Verify accordion expands by toggle click', async () => {
+      await page.locator('[data-ui-name="ButtonLink"]').first().click();
+      const accordionRow = page.locator('[aria-rowindex="5"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #0']);
+    });
+
+    await test.step('Verify accordion expands cell click', async () => {
+      messages = [];
+      const row = page.locator('[aria-rowindex="6"]');
+      await row.locator('[data-ui-name="Body.Cell"][aria-colindex="2"]').first().click();
+      const accordionRow = page.locator('[aria-rowindex="9"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #1']);
+    });
+
+    await test.step('Verify accordion collapses by toggle click', async () => {
+      messages = [];
+      await page.locator('[data-ui-name="ButtonLink"]').first().click();
+      const accordionRow = page.locator('[aria-rowindex="5"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'hidden' });
+      await page.waitForEvent('console', {
+        predicate: (msg) => msg.type() === 'log' && msg.text() === 'Accordion close for row #0',
+        timeout: 200,
+      });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion close for row #0']);
+    });
+
+    await test.step('Verify accordion collapses cell click', async () => {
+      messages = [];
+      const row = page.locator('[aria-rowindex="3"]');
+      await row.locator('[data-ui-name="Body.Cell"][aria-colindex="1"]').first().click();
+
+      const accordionRow = page.locator('[aria-rowindex="6"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'hidden' });
+      await page.waitForEvent('console', {
+        predicate: (msg) => msg.type() === 'log' && msg.text() === 'Accordion close for row #1',
+        timeout: 200,
+      });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion close for row #1']);
+    });
+  });
+
+  test('Verify table in table keyboard navigation when accordionMode=toggle', async ({ page }) => {
+    let messages: string[] = [];
+    page.on('console', (msg) => {
+      if (msg.type() === 'log') {
+        messages.push(msg.text());
+      }
+    });
+    const standPath = 'stories/components/data-table/docs/examples/table-in-table.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en', {
+      accordionMode: 'toggle',
+      onAccordionToggle: 'function(type, rowIndex) { console.log("Accordion " + type + " for row #" + rowIndex); }',
+    });
+    await page.setContent(htmlContent);
+
+    const rows = page.locator('[data-ui-name="Body.Row"]');
+    await page.keyboard.press('Tab');
+
+    await test.step('Verify it is possible to scroll the last cell by keyboard', async () => {
+      const rowCount = await rows.count();
+
+      for (let i = 0; i < rowCount; i++) {
+        await page.keyboard.press('ArrowDown');
+      }
+      const lastRow = page.locator('[data-ui-name="Body.Row"][aria-rowindex="6"]');
+      const cellinLastRow = lastRow.locator(
+        '[data-ui-name="Body.Cell"][aria-colindex="1"][data-aria-level="1"]',
+      );
+
+      await expect(cellinLastRow).toBeFocused();
+    });
+
+    await test.step('Verify keyboard navigation in table works well when one accordion axpanded', async () => {
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('Enter');
+      const accordionRow = page.locator('[aria-rowindex="5"][aria-level="2"]');
+
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #0']);
+      for (let i = 0; i < 7; i++) await page.keyboard.press('ArrowDown');
+      const lastRow = page.locator('[data-ui-name="Body.Row"][aria-rowindex="9"]');
+      const cellinLastRow = lastRow.locator(
+        '[data-ui-name="Body.Cell"][aria-colindex="1"][data-aria-level="1"]',
+      );
+
+      await expect(cellinLastRow).toBeFocused();
+    });
+
+    await test.step('Verofy forst accordion closed when second is collapsed and keyboard navigation not broken', async () => {
+      messages = [];
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('Enter');
+      const accordionRow = page.locator('[aria-rowindex="6"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(2);
+      expect(messages).toEqual(['Accordion open for row #1',
+        'Accordion close for row #0']);
       for (let i = 0; i < 6; i++) await page.keyboard.press('ArrowDown');
 
       const lastRow = page.locator('[data-ui-name="Body.Row"][aria-rowindex="9"]');
@@ -394,6 +593,67 @@ test.describe('Accordion in table', () => {
       );
 
       await expect(cellinLastRow).toBeFocused();
+    });
+
+    await test.step('Verify keyboard navigation not broken when second accordion just expanded', async () => {
+      messages = [];
+
+      for (let i = 0; i < 6; i++) await page.keyboard.press('ArrowUp');
+      await page.keyboard.press('Enter');
+      const accordionRow = page.locator('[aria-rowindex="6"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'hidden' });
+      await page.waitForEvent('console', {
+        predicate: (msg) => msg.type() === 'log' && msg.text() === 'Accordion close for row #1',
+        timeout: 200,
+      });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion close for row #1']);
+      messages = [];
+      await page.keyboard.press('Enter');
+      await accordionRow.waitFor({ state: 'visible' });
+
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #1']);
+
+      await page.keyboard.press('ArrowUp');
+      await expect(page.locator('[data-ui-name="ButtonLink"]').first()).toBeFocused();
+    });
+  });
+
+  test('Verify table in table mouse navigation when accordionMode=toggle', async ({ page }) => {
+    let messages: string[] = [];
+    page.on('console', (msg) => {
+      if (msg.type() === 'log') {
+        messages.push(msg.text());
+      }
+    });
+
+    const standPath = 'stories/components/data-table/docs/examples/table-in-table.tsx';
+
+    const htmlContent = await e2eStandToHtml(standPath, 'en', {
+      accordionMode: 'toggle',
+      onAccordionToggle: 'function(type, rowIndex) { console.log("Accordion " + type + " for row #" + rowIndex); }',
+    });
+
+    await page.setContent(htmlContent);
+
+    await test.step('Verify accordion expands by toggle click', async () => {
+      await page.locator('[data-ui-name="ButtonLink"]').first().click();
+      const accordionRow = page.locator('[aria-rowindex="5"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+      expect(messages.length).toBe(1);
+      expect(messages).toEqual(['Accordion open for row #0']);
+    });
+
+    await test.step('Verify accordion expands cell click and prev accordion closed', async () => {
+      messages = [];
+      const row = page.locator('[aria-rowindex="6"]');
+      await row.locator('[data-ui-name="Body.Cell"][aria-colindex="2"]').first().click();
+      const accordionRow = page.locator('[aria-rowindex="6"][aria-level="2"]');
+      await accordionRow.waitFor({ state: 'visible' });
+
+      expect(messages.length).toBe(2);
+      expect(messages).toEqual(['Accordion open for row #1', 'Accordion close for row #0']);
     });
   });
 
@@ -757,5 +1017,25 @@ test.describe('Accordion in table', () => {
       await expect(sortIconKeywordAcc).toHaveAttribute('aria-expanded', 'false');
       await expect(datatableChild).not.toBeVisible();
     });
+  });
+
+  test('Verify mouse interaction on a cell with merged rows', async ({ page }) => {
+    const standPath =
+      'stories/components/data-table/advanced/examples/accordion_in_merged_rows.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    const collapse = page.locator('[data-ui-name="Collapse"]');
+    const spannedCell = page.locator('[data-ui-name="Body.Row"] div[aria-rowspan="3"]');
+    const showDetails = page.locator('[aria-label="Show details"]').first();
+
+    await expect(collapse).toBeHidden();
+
+    await spannedCell.click();
+    await expect(collapse).toBeVisible();
+
+    await showDetails.click();
+    await expect(collapse).toBeHidden();
   });
 });